refactor(cards): updated card design and refactored code for reuse.
