+++ src/app/globals.css
@@ -20,20 +20,12 @@ body {
   place-items: center;
   min-width: 320px;
   min-height: 100vh;
+  color: white;
 }
 
 h1 {
   font-size: 3.2em;
-  line-height: 1.1;
-  color: white;
-}
-
-h2 {
-  color: white;
-}
-
-p {
-  color: white;
+  line-height: 1;
 }
 
 a:link, a:visited{
