+++ src/app/launches/components/LaunchCard/index.tsx
@@ -7,24 +7,16 @@ import React from 'react'
 // COMPONENTS
 import Overlay from '../../../../components/Overlay'
 import LaunchOverlay from '../LaunchOverlay'
+import Card from '../../../../components/Card'
 
 // TYPES
-import { Launch } from '../../../../types/launches/Launch'
-
-// Next
-import Image from 'next/image'
-
-// Images
-import spacex_logo from '../../../../assets/spacex_default_logo.png'
+import { LaunchCardType } from '../../../../types/launches'
 
 // CSS
 import './index.css'
+import parseDate from '../../../../utility/date/parseDate'
 
-interface LaunchCardProps {
-    item: Launch
-}
-
-const LaunchCard: React.FC<LaunchCardProps> = ({ item }) => {
+const LaunchCard: React.FC<LaunchCardType> = ({ item }) => {
     // Handles overlay visibility
     const [isOverlayVisible, setIsOverlayVisible] = useState(false)
 
@@ -36,14 +28,11 @@ const LaunchCard: React.FC<LaunchCardProps> = ({ item }) => {
         setIsOverlayVisible(false)
     }
 
+    const date: string = parseDate(item.date_utc)
+
     return (
         <>
-            <div className="card">
-                <Image src={item.links.patch.large || spacex_logo} alt={item.name} width={300} height={300} />
-                <h2>{item.name}</h2>
-                <p>{item.date_utc}</p>
-                <button onClick={handleDisplayOverlay}>View Details</button>
-            </div>
+            <Card title={item.name} subtitle={"Launch Date:" + <br /> + `${date} UTC`} image={item.links.patch.large} imageAlt={item.name} handleDisplayOverlay={handleDisplayOverlay} />
             {/* Makes sure we are using the Overlay styling which is globally scoped */}
             {isOverlayVisible && <Overlay item={item} ItemComponent={LaunchOverlay} onClose={handleHideOverlay} />}
         </>
