+++ src/components/Card/index.css
@@ -1,0 +1,27 @@
+.card {
+    background-color: rgba(255, 255, 255, .1);
+    box-shadow: 0 25px 45px rgba(0, 0, 0, .2);
+    border-radius: 10px;
+    border: 2px solid rgba(255, 255, 255, .5);
+    backdrop-filter: blur(5px);
+    border-bottom: 1px solid rgba(255, 255, 255, .2);
+    border-right: 1px solid rgba(255, 255, 255, .2);
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.card__image {
+    width: 100%;
+}
+
+.card__title {
+    font-size: 1.5rem;
+    color: white;
+}
+
+.card__subtitle {
+    font-size: 1rem;
+    color: white;
+}
