+++ src/components/CardGallery/index.css
@@ -4,4 +4,5 @@
     justify-content: center;
     margin-top: 8rem;
     margin-bottom: 8rem;
+    gap: 2rem;
 }
