+++ src/components/Header/index.tsx
@@ -10,7 +10,7 @@ import './header.min.css'
 const Header: React.FC = () => {
     return (
         <header>
-            <h1>G/R</h1>
+            <h1 className="header__h1">G/R</h1>
             <Navbar />
         </header>
     )
