+++ src/types/cards/index.ts
@@ -1,0 +1,7 @@
+export interface CardType {
+    title: string;
+    subtitle: string;
+    image: string;
+    imageAlt: string;
+    handleDisplayOverlay: () => void;
+}
