+++ src/utility/date/parseDate.ts
@@ -1,0 +1,4 @@
+
+export default function parseDate(date: string): string {
+    return new Date(date).toLocaleString('en-US', { timeZone: 'UTC' });
+}
